+++ src/index.tsx
@@ -1,10 +1,13 @@
 import React from 'react'
 import { registerRootComponent } from 'expo'
+import { NavigationContainer  } from '@react-navigation/native'
 
 import App from './App'
 
 const AppContainer = () => (
-  <App />
+  <NavigationContainer>
+    <App />
+  </NavigationContainer>
 )
 
 export default registerRootComponent(AppContainer)