+++ src/interfaces/index.ts
@@ -1,0 +1,7 @@
+interface ICreateUserState {
+  name : string
+  email: string
+  phone: string
+}
+
+export { ICreateUserState }