+++ changelog.md
@@ -1,4 +1,9 @@
 # React Native Firebase App with Expo
 
+## Version 0.2.0
+  - Implemented configuration to work with screens.
+  - Implemented CreateUser screen, with its UI and its logic.
+  - Implemented UserDetail and UsersList screens, but files only.
+
 ## Version 0.1.0
   - Implemented project configuration.